Move expand function to base class
