+++ src/aspire/basis/ffb_3d.py
@@ -1,10 +1,8 @@
 import logging
 import numpy as np
 from numpy import pi
-from scipy.sparse.linalg import LinearOperator, cg
 
 from aspire.nfft import anufft3, nufft3
-from aspire.utils import ensure
 from aspire.utils.matrix import roll_dim, unroll_dim
 from aspire.utils.matlab_compat import m_flatten, m_reshape
 from aspire.basis.basis_utils import sph_bessel, norm_assoc_legendre, lgwt
@@ -337,50 +335,5 @@ class FFBBasis3D(FBBasis3D):
             # TODO: Fix this to avoid lookup each time.
             ind = self._indices['ells'] == ell
             v[ind, :] = v_ell
-        v = roll_dim(v, sz_roll)
-        return v
-
-    def expand(self, x):
-
-        """
-        Obtain expansion coefficients in FB basis from those in standard 3D coordinate basis
-
-        This is a similar function to evaluate_t but with more accuracy by using
-        the cg optimizing of linear equation, Ax=b.
-
-        :param x: An array whose first three dimensions are to be expanded in FB basis.
-             These dimensions must equal `self.sz`.
-        :return : The coefficients of `v` expanded in FB basis. The first dimension
-            of `v` is with size of `count` and the second and higher dimensions
-            of the return value correspond to those higher dimensions of `x`.
-
-        """
-        # TODO: this is function could be move to base class if all standard and fast versions of 2d and 3d are using
-        #       the same data structures of x and v.
-        # ensure the first three dimensions with size of self.sz
-        x, sz_roll = unroll_dim(x, self.ndim + 1)
-        x = m_reshape(x, (self.sz[0], self.sz[1], self.sz[2], -1))
-
-        ensure(x.shape[:self.ndim] == self.sz, f'First {self.ndim} dimensions of x must match {self.sz}.')
-
-        operator = LinearOperator(shape=(self.count, self.count),
-                                  matvec=lambda v: self.evaluate_t(self.evaluate(v)))
-
-        # TODO: (from MATLAB implementation) - Check that this tolerance make sense for multiple columns in v
-        tol = 10*np.finfo(x.dtype).eps
-        logger.info('Expanding array in basis')
-
-        # number of image samples
-        n_data = np.size(x, self.ndim)
-        v = np.zeros((self.count, n_data), dtype=x.dtype)
-
-        for isample in range(0, n_data):
-            b = self.evaluate_t(x[..., isample])
-            # TODO: need check the initial condition x0 can improve the results or not.
-            v[..., isample], info = cg(operator, b, tol=tol)
-            if info != 0:
-                raise RuntimeError('Unable to converge!')
-
-        # return v coefficients with the first dimension of self.count
         v = roll_dim(v, sz_roll)
         return v
